Put run task for Scala.js in ScalaJsToolchain
